changes server.js name
